requirements.txt
google-adk
python-dotenv==1.1.0
google-genai